rasberry-pi-4-server-firebase-no-sql-wen-cam-pagekite/requirements.txt
# Fish Feeder Pi Server Requirements

# ===== CORE DEPENDENCIES =====
flask==2.3.3
flask-cors==4.0.0
pyserial==3.5
requests==2.31.0

# ===== HIGH-PERFORMANCE JSON =====
orjson==3.9.10

# ===== FIREBASE & GOOGLE CLOUD =====
firebase-admin==6.2.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-cloud-firestore==2.11.1

# ===== COMPUTER VISION =====
opencv-python==4.8.1.78

# ===== WEBSOCKET & REAL-TIME =====
python-socketio==5.10.0
eventlet==0.33.3

# ===== SYSTEM MONITORING =====
psutil==5.9.5

# ===== DATA PROCESSING =====
pandas==2.0.3
numpy==1.24.3

# ===== UTILITIES =====
python-dotenv==1.0.0
schedule==1.2.0
protobuf==4.25.1
watchdog==3.0.0
rich==13.7.0

# ===== DEVELOPMENT TOOLS (Optional) =====
pytest==7.4.2
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
mypy==1.8.0

# ===== PERFORMANCE MONITORING (Optional) =====
memory-profiler==0.61.0
line-profiler==4.1.1 